# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Billing & Solar ERP API",
        "description": "Billing + Cloud Solar Panel Installation & Service Management API. Workflow: Lead → Survey → Quotation → Installation → Service & AMC. Roles: MD Super Admin, Branch Manager, Branch Admin, Sales Executive, Site Survey Engineer, Installation Manager/Technician, Service Team, Accounts. JWT Secured: use POST /api/v1/auth/login then header Authorization: Bearer {access_token}. AI-friendly: structured JSON responses, clear field names, dashboard metrics.",
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "API Server (localhost)"
        },
        {
            "url": "https://testbillapi.eazycutz.com/api/v1",
            "description": "API Server (Live)"
        },
        {
            "url": "/api/v1",
            "description": "API Server (relative)"
        }
    ],
    "paths": {
...
        "/solar/vendors": {
            "get": {
                "tags": [
                    "Solar - Vendors"
                ],
                "summary": "List Vendors",
                "description": "List vendors (branch or all for MD Super Admin).",
                "operationId": "2b81257a9bcd07d2de0bf016db5533db",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Solar - Vendors"
                ],
                "summary": "Create Vendor",
                "description": "Create vendor (MD Super Admin / Branch).",
                "operationId": "88b8f1ee7e11127c0be0affa2925fc84",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "contact_person": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "object"
                                    },
                                    "gstin": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solar/vendors/{id}": {
            "get": {
                "tags": [
                    "Solar - Vendors"
                ],
                "summary": "Get Vendor",
                "description": "Get vendor.",
                "operationId": "a120ce1c68e9ef162928948d9040b8e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Solar - Vendors"
                ],
                "summary": "Update Vendor",
                "description": "Update vendor.",
                "operationId": "1ca797d0c74b24b028f073f5d5a381c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "contact_person": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "object"
                                    },
                                    "gstin": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Solar - Vendors"
                ],
                "summary": "Delete Vendor",
                "description": "Delete vendor.",
                "operationId": "894fe626baca1f997436eab3506138f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: \\\"Authorization: Bearer {token}\\\"",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}